refactor code, address part of the reviews.
